Add GVL hold time tracking
Attempts to add a gvl hold time value to both the Global and Local
tracking information.

The Global value seems to be tracking. However the Local value does not.
My current suspicion is that the event for
RUBY_INTERNAL_THREAD_EVENT_SUSPENDED is not called in the same thread
context as the other callbacks and thus the wrong thread-local variable
is being updated. The global works since it is a shared resource across
all threads.