Fix compatibility with Ruby 3.3
There was a number of changes in the GVL instrumentation API in
3.3, most notably events are no longer guaranteed to be triggered
from the concerned native thread, so we can no longer rely on
C thread local.

Instead we receive the concerned thread as callback argument,
and we can use the rb_internal_thread_specific_* API to store
thread local data.